+++ CHANGELOG.md
@@ -27,10 +27,10 @@ This is an example of a brief overview of the _Major_ or _Minor_ version changes
 
 ---
 
-## [6.5] `border-focus`
+## [6.5] `border--focus`
 
-- Added `border-focus` and `border-focus--no-trsn' helper classes for standard focus styles
-- Added `trsn-default` helper class for default transition
+- Added `border--focus` and `border--focus--no-trsn' helper classes for standard focus styles
+- Added `transition-default` helper class for default transition
 
 ## [6.4] `Avatar`